respond to deserialization errors with json
If there is an error while deserializing, it is nicer to respond with a
JSON response, rather than plain text. Use a custom Deserialize action
class, which puts the error into the stash where either the Serialize
action class or the JSON view expects. And then skip the rest of the
request.